Various changes
Edit review working correctly. Remember me option in log in. Fixing bug of profile-info page. Some style changes

+++ paw-webapp/interfaces/src/main/java/ar/edu/itba/paw/persistance/ContentDao.java
@@ -15,6 +15,6 @@ public interface ContentDao {
     List<Content> getSearchedContent(String query);
     List<Content> getSearchedContentRandom(String query);
     List<Content> findByType(String type);
-    void addContentPoints(int contentId,int rating);
+    void addContentPoints(long contentId,int rating);
     void decreaseContentPoints(long contentId,int rating);
 }
