+++ paw-webapp/persistance/src/main/java/ar/edu/itba/paw/persistance/ContentJdbcDao.java
@@ -207,7 +207,7 @@ public class ContentJdbcDao implements ContentDao {
     }
 
     @Override
-    public void addContentPoints(int contentId,int rating){
+    public void addContentPoints(long contentId,int rating){
         template.update("UPDATE content SET rating = rating + ?, reviewsAmount = reviewsAmount+1 WHERE id = ?", new Object[] {rating, contentId});
     }
     @Override
